+++ packages/frontend/src/components/SimpleTable/CellContextMenu.tsx
@@ -2,6 +2,7 @@ import { Menu, MenuDivider } from '@blueprintjs/core';
 import { MenuItem2 } from '@blueprintjs/popover2';
 import { subject } from '@casl/ability';
 import { isDimension, isField, ResultRow } from '@lightdash/common';
+import mapValues from 'lodash-es/mapValues';
 import React, { FC } from 'react';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import { useParams } from 'react-router-dom';
@@ -16,16 +17,18 @@ import DrillDownMenuItem from '../MetricQueryData/DrillDownMenuItem';
 import { useMetricQueryDataContext } from '../MetricQueryData/MetricQueryDataProvider';
 
 const CellContextMenu: FC<Pick<CellContextMenuProps, 'cell'>> = ({ cell }) => {
-    const { openUnderlyingDataModel } = useMetricQueryDataContext();
+    const { openUnderlyingDataModal } = useMetricQueryDataContext();
     const { showToastSuccess } = useToaster();
     const meta = cell.column.columnDef.meta;
     const item = meta?.item;
 
     const value: ResultRow[0]['value'] = cell.getValue()?.value || {};
+    const fieldValues = mapValues(cell.row.original, (v) => v?.value) || {};
 
     const { track } = useTracking();
     const { user } = useApp();
     const { projectUuid } = useParams<{ projectUuid: string }>();
+
     return (
         <Menu>
             {item && value.raw && isField(item) && (
@@ -55,11 +58,11 @@ const CellContextMenu: FC<Pick<CellContextMenuProps, 'cell'>> = ({ cell }) => {
                         text="View underlying data"
                         icon="layers"
                         onClick={() => {
-                            openUnderlyingDataModel(
+                            openUnderlyingDataModal({
+                                item: meta.item,
                                 value,
-                                meta,
-                                cell.row.original || {},
-                            );
+                                fieldValues,
+                            });
                             track({
                                 name: EventName.VIEW_UNDERLYING_DATA_CLICKED,
                                 properties: {
@@ -81,9 +84,9 @@ const CellContextMenu: FC<Pick<CellContextMenuProps, 'cell'>> = ({ cell }) => {
                 })}
             >
                 <DrillDownMenuItem
-                    row={cell.row.original || {}}
+                    item={item}
+                    fieldValues={fieldValues}
                     pivotReference={meta?.pivotReference}
-                    selectedItem={item}
                     trackingData={{
                         organizationId: user?.data?.organizationUuid,
                         userId: user?.data?.userUuid,